+++ cobaya/component.py
@@ -12,7 +12,7 @@ from cobaya.typing import Any, InfoDict, InfoDictIn, empty_dict
 from cobaya.tools import resolve_packages_path, load_module, get_base_classes, \
     get_internal_class_component_name, deepcopy_where_possible, VersionCheckError
 from cobaya.conventions import kinds, cobaya_package, reserved_attributes
-from cobaya.yaml import yaml_load_file, yaml_dump
+from cobaya.yaml import yaml_load_file, yaml_dump, yaml_load
 from cobaya.mpi import is_main_process
 
 
@@ -167,8 +167,7 @@ class HasDefaults:
         from this class, it will return the result from an inherited class if that
         provides bibtex.
         """
-        filename = cls.__dict__.get('bibtex_file')
-        if filename:
+        if filename := cls.__dict__.get('bibtex_file'):
             bib = cls.get_text_file_content(filename)
         else:
             bib = cls.get_associated_file_content('.bibtex')
@@ -271,12 +270,16 @@ class HasDefaults:
         yaml_text = cls.get_associated_file_content('.yaml')
         options = cls.get_class_options(input_options=input_options)
         if options and yaml_text:
-            log = get_logger(cls.get_qualified_class_name())
-            raise LoggedError(log,
-                              "%s: any class can either have .yaml or class variables "
-                              "but not both (type declarations without values are fine "
-                              "also with yaml file). You have class attributes: %s",
-                              cls.get_qualified_class_name(), list(options))
+            yaml_options = yaml_load(yaml_text)
+            if both := set(yaml_options).intersection(options):
+                raise LoggedError(get_logger(cls.get_qualified_class_name()),
+                                  "%s: class has .yaml and class variables/options "
+                                  "that define the same keys: %s \n"
+                                  "(type declarations without values are fine "
+                                  "with yaml file as well).",
+                                  cls.get_qualified_class_name(), list(both))
+            options |= yaml_options
+            yaml_text = None
         if return_yaml and not yaml_expand_defaults:
             return yaml_text or ""
         this_defaults = yaml_load_file(cls.get_yaml_file(), yaml_text) \
@@ -293,6 +296,17 @@ class HasDefaults:
             return yaml_dump(defaults)
         else:
             return defaults
+
+    # noinspection PyUnusedLocal
+    @classmethod
+    def get_modified_defaults(cls, defaults, input_options=empty_dict):
+        """
+        After defaults dictionary is loaded, you can dynamically modify them here
+        as needed,e.g. to add or remove defaults['params']. Use this when you don't
+        want the inheritance-recursive nature of get_defaults() or don't only
+        want to affect class attributes (like get_class_options() does0.
+        """
+        return defaults
 
     @classmethod
     def get_annotations(cls) -> InfoDict:
@@ -325,6 +339,7 @@ class CobayaComponent(HasLogger, HasDefaults):
         if standalone:
             # TODO: would probably be more natural if defaults were always read here
             default_info = self.get_defaults(input_options=info)
+            default_info = self.get_modified_defaults(default_info, input_options=info)
             default_info.update(info)
             info = default_info
 
@@ -455,8 +470,7 @@ class ComponentCollection(dict, HasLogger):
         self[name] = component
 
     def dump_timing(self):
-        timers = [component for component in self.values() if component.timer]
-        if timers:
+        if timers := [component for component in self.values() if component.timer]:
             sep = "\n   "
             self.log.info(
                 "Average computation time:" + sep + sep.join(