+++ cobaya/theory.py
@@ -177,8 +177,8 @@ class Theory(CobayaComponent):
 
         :return: iterable of parameter names
         """
-        params = getattr(self, "params", None)
-        if params:
+
+        if params := getattr(self, "params", None):
             return [k for k, v in params.items() if
                     hasattr(v, 'get') and v.get('derived') is True]
         else:
@@ -378,7 +378,8 @@ class TheoryCollection(ComponentCollection):
                             logger=self.log, component_path=info.get("python_path"))
                     self.add_instance(
                         name, theory_class(
-                            info, packages_path=packages_path, timing=timing, name=name))
+                            info, packages_path=packages_path, timing=timing, name=name,
+                            standalone=False))
 
     def __getattribute__(self, name):
         if not name.startswith('_'):